rename to deployerCfg and updates
